Revert unneeded fix for sending JS events in Android
